Another refactor into smaller, clearly labeled modules
