+++ pyGr/zoning_algorithm/octtree.py
@@ -1,6 +1,7 @@
 from Queue import Queue
 from helper_functions import *
 import numpy as np
+from pyGr.common.region_ops import get_region_boundary
 
 class Octtree:
     fid_counter = 0
